Cmm switch on two alts + def uses two equality tests instead of range checks.
Summary:
For code of the likes of:
f 1 = e1
f 7 = e2
f _ = e3

Ghc first checks for x>7, then checks for <= 7 and at the end checks for
equality on 1.

This is both larger codewise and at least on my Intel 6700K slower than
simply checking for equality twice in most cases.

For arguments successfully predicted to be >7 it's still not measurably
faster than the if/else if/else variant.

This is WIP related to #14644

Reviewers: bgamari, simonmar

Subscribers: rwbarton, thomie, carter

Differential Revision: https://phabricator.haskell.org/D4294