Start over with #14691 and only add EvExpr to EvTerm
this can be useful for typechecker plugins, without disrupting the
architecture of where Typeable evidence is created in the desgurar, not
the typechecker.

Some of the non-recursive constructors, such as `EvDelayedError` or
`EvLit`, could possibly be dropped in favor of storing the `CoreExpr`
directory. But there is not much to be gained from that.