Update JMP_TBL targets during shortcutting in X86 NCG.
Without updating the JMP_TBL information the block list in
JMP_TBL contained references which were eliminated by shortcutting.

The actual assembly generation doesn't look at these fields so this
didn't cause any bugs yet. However as long as we carry this information
around we should make an effort to keep it correct.

Especially since it's useful for debugging purposes and can be used
for passes near the end of the codegen pipeline.