Refactor some aspects of the layout algorithm and code.
Don't completely ignore calls.

Laying out a call and it's return address sequentially should
still be a win IF it doesn't infer with any other aspect of
code layout. So we take calls into consideration but only for
the last pass where we combine neighbouring chains.

Also make buildChains non-recursive and use a go-style function
for the recursion instead along with some renaming.